# docs/openapi/agent-access-v1.yaml
openapi: 3.0.3
info:
  title: ActWeave Agent Access Protocol
  version: '2026-07-20'
  description: >-
    Workspace- and Agent-scoped API for service principals. Management user JWTs
    are not valid on this surface. Persisted Run events use the AAP v1 Event
    Envelope; SSE transport signals are not persisted events.
servers:
  - url: /api/agent-access/v1
security:
  - AAPAccessToken: []
paths:
...
  /workspaces/{workspaceId}/agents/{agentId}/files:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/AgentId'}
    post:
      operationId: createFile
      summary: Create an AAP file upload intent (presigned PUT).
      description: >-
        Allocates a fileId and returns a short-lived presigned PUT URL with
        Content-Length and Content-Type bound in the signature. The upload
        object appears only on this response — subsequent GET never echoes it.
        Requires file:write. Gated by agentAccess.files (default disabled).
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateFileRequest'}
      responses:
        '201':
          description: Upload intent created.
          headers:
            Location: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CreateFileResponse'}
        '200':
          description: Idempotent replay of the same create intent.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CreateFileResponse'}
        '404': {$ref: '#/components/responses/NotVisible'}
        '409': {$ref: '#/components/responses/IdempotencyConflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /workspaces/{workspaceId}/agents/{agentId}/files/{fileId}:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/AgentId'}
      - {$ref: '#/components/parameters/FileId'}
    get:
      operationId: getFile
      summary: Get file status (source of truth; no SSE in v1).
      description: >-
        Returns file lifecycle status, processing stages, and a relative content
        link. Never includes upload URL, presign headers, or live download URLs.
        Requires file:read.
      responses:
        '200':
          description: File resource.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GetFileResponse'}
        '404': {$ref: '#/components/responses/NotVisible'}
  /workspaces/{workspaceId}/agents/{agentId}/files/{fileId}:complete:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/AgentId'}
      - {$ref: '#/components/parameters/FileId'}
    post:
      operationId: completeFile
      summary: Confirm staging upload (fast path; does not wait for promote).
      description: >-
        Stats staging, validates size/MIME, CAS to UPLOADED, enqueues promote.
        Requires file:write and ownership. Does not block on encryption.
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                sha256: {type: string, pattern: '^[a-fA-F0-9]{64}$'}
      responses:
        '200':
          description: Complete accepted; promote is asynchronous.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CompleteFileResponse'}
        '404': {$ref: '#/components/responses/NotVisible'}
        '409': {$ref: '#/components/responses/IdempotencyConflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /workspaces/{workspaceId}/agents/{agentId}/files/{fileId}/content:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/AgentId'}
      - {$ref: '#/components/parameters/FileId'}
    get:
      operationId: getFileContent
      summary: Stream file content with Bearer (path A).
      description: >-
        Server-side SecureStore.Open stream. Requires file:read and READY file.
        Prefer :download for large bodies.
      responses:
        '200':
          description: Decrypted file body.
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/NotVisible'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/RateLimited'}
...
  /files/downloads/{tokenId}:
    parameters:
      - {$ref: '#/components/parameters/TokenId'}
    get:
      operationId: downloadFileByToken
      summary: Stream file content via opaque download token (no Bearer).
      security: []
      description: >-
        Validates the opaque token row (expiry, purpose, not consumed, file READY)
        and streams SecureStore body. Token is not a JWT and is not a MinIO key.
      responses:
        '200':
          description: Decrypted file body.
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/NotVisible'}
components:
  securitySchemes:
    AAPAccessToken:
      type: http
      scheme: bearer
      bearerFormat: AAP JWT
      description: Short-lived token with Agent Access audience; user JWTs are rejected.
  parameters:
    WorkspaceId: {name: workspaceId, in: path, required: true, schema: {$ref: '#/components/schemas/UUID'}}
    AgentId: {name: agentId, in: path, required: true, schema: {$ref: '#/components/schemas/UUID'}}
...
    FileId: {name: fileId, in: path, required: true, schema: {$ref: '#/components/schemas/UUID'}}
    TokenId: {name: tokenId, in: path, required: true, schema: {$ref: '#/components/schemas/UUID'}}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
...
  schemas:
    UUID: {type: string, format: uuid}
...
    CreateFileRequest:
      type: object
      additionalProperties: false
      required: [mediaType, sizeBytes]
      properties:
        filename: {type: string, maxLength: 512}
        mediaType:
          type: string
          enum: [image/png, image/jpeg, image/webp, image/gif, application/pdf]
        sizeBytes: {type: integer, format: int64, minimum: 1}
        sha256: {type: string, pattern: '^[a-fA-F0-9]{64}$'}
        purpose:
          type: string
          enum: [GENERAL, VISION, DOCUMENT, TOOL_INPUT]
    File:
      type: object
      additionalProperties: false
      required: [object, id, agentId, status, mediaType, sizeBytes, purpose, processing, artifacts, links, createdAt, updatedAt]
      description: >-
        Public File resource. Must never include upload, presign, or live
        downloadUrl fields (sensitive acceptance).
      properties:
        object: {type: string, enum: [file]}
        id: {$ref: '#/components/schemas/UUID'}
        agentId: {$ref: '#/components/schemas/UUID'}
        status:
          type: string
          enum: [pending_upload, uploaded, processing, ready, failed, expired]
        filename: {type: string}
        mediaType: {type: string}
        detectedMediaType: {type: string}
        sizeBytes: {type: integer, format: int64}
        sha256: {type: string}
        purpose: {type: string}
        error: {$ref: '#/components/schemas/ErrorValue'}
        processing:
          type: object
          additionalProperties: false
          required: [version, stages]
          properties:
            version: {type: integer, format: int64}
            stages:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [stage, status]
                properties:
                  stage: {type: string}
                  status: {type: string}
        artifacts: {type: array, items: {type: object}}
        links:
          type: object
          additionalProperties: false
          required: [content]
          properties:
            content: {type: string, description: Relative Bearer content path (not a live token URL).}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        readyAt: {type: string, format: date-time}
    FileUpload:
      type: object
      additionalProperties: false
      required: [method, url, headers, expiresAt]
      description: Write-only create response fragment; never appears on GET.
      properties:
        method: {type: string, enum: [PUT]}
        url: {type: string, format: uri}
        headers:
          type: object
          additionalProperties: {type: string}
          description: Must include Content-Type and Content-Length for signed PUT.
        expiresAt: {type: string, format: date-time}
    CreateFileResponse:
      type: object
      additionalProperties: false
      required: [file, idempotent]
      properties:
        file: {$ref: '#/components/schemas/File'}
        upload: {$ref: '#/components/schemas/FileUpload'}
        idempotent: {type: boolean}
    CompleteFileResponse:
      type: object
      additionalProperties: false
      required: [file, idempotent]
      properties:
        file: {$ref: '#/components/schemas/File'}
        idempotent: {type: boolean}
    GetFileResponse:
      type: object
      additionalProperties: false
      required: [file]
      properties:
        file: {$ref: '#/components/schemas/File'}
...
    ErrorValue:
      type: object
      additionalProperties: false
      required: [code, message, retryable]
      properties:
        code: {type: string}
        message: {type: string}
        retryable: {type: boolean}
        details: {type: object}
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/ErrorValue'}
  responses:
...
    NotVisible:
      description: Resource does not exist or is not visible in the bound Scope.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    ValidationError:
      description: Request validation failed.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
    IdempotencyConflict:
      description: The key was already used with a different normalized request hash.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
    RateLimited:
      description: One or more Workspace, Client, Agent, or Subject quota buckets is exhausted.
      headers:
        Retry-After: {schema: {type: integer, minimum: 1}}
        RateLimit-Limit: {schema: {type: integer}}
        RateLimit-Remaining: {schema: {type: integer}}
        RateLimit-Reset: {schema: {type: integer}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}